+++ src/generator/gen_common.rs
@@ -365,12 +365,13 @@ impl GeneratorCore {
     }
 
     /// Get a field name with proper casing
-    pub fn get_field_name(&self, reg: &RifRegInst, field: &RifFieldInst) -> String {
+    pub fn get_field_name(&self, reg: &RifRegInst, field: &RifFieldInst, field_array: bool) -> String {
         // println!("[get_field_name] {}.{} : rsvd={}, field array = {:?}, reg array={:?}",
         //     reg.reg_name, f.name, f.is_reserved(), f.array, reg.array);
+        let flat_name = !field_array && (field.array.dim() > 1 || reg.array.dim()==0 || reg.array.is_inst());
         if field.is_reserved() && self.setting.privacy.is_public() {
             format!("rsvd{}",field.lsb)
-        } else if field.array.dim() > 1 || reg.array.dim()==0 || reg.array.is_inst() {
+        } else if flat_name {
             field.name_flat().to_casing(self.setting.casing)
         } else {
             field.name.to_casing(self.setting.casing)